move declare_experts and get_experts to dht_ops.py
