+++ tests/benchmark_dht.py
@@ -5,6 +5,8 @@ import time
 from tqdm import trange
 
 import hivemind
+import hivemind.client.expert_uid
+import hivemind.client.dht_ops
 from hivemind.utils.threading import increase_file_limit
 
 logger = hivemind.get_logger(__name__)
@@ -42,7 +44,7 @@ def benchmark_dht(num_peers: int, initial_peers: int, num_experts: int, expert_b
     for start in trange(0, num_experts, expert_batch_size):
         store_start = time.perf_counter()
         endpoints.append(random_endpoint())
-        successes = store_peer.declare_experts(expert_uids[start: start + expert_batch_size], endpoints[-1]).values()
+        successes = hivemind.client.dht_ops.declare_experts(expert_uids[start: start + expert_batch_size], endpoints[-1]).values()
         total_store_time += time.perf_counter() - store_start
 
         total_stores += len(successes)
@@ -60,7 +62,7 @@ def benchmark_dht(num_peers: int, initial_peers: int, num_experts: int, expert_b
 
     for start in trange(0, len(expert_uids), expert_batch_size):
         get_start = time.perf_counter()
-        get_result = get_peer.get_experts(expert_uids[start: start + expert_batch_size])
+        get_result = hivemind.client.dht_ops.get_experts(expert_uids[start: start + expert_batch_size])
         total_get_time += time.perf_counter() - get_start
 
         for i, expert in enumerate(get_result):
